requirements.txt
numpy==1.24.4
scipy==1.10.1
scikit-learn==1.3.2
matplotlib==3.7.5
pandas==1.4.1
wandb==0.18.5
matplotlib==3.7.5
tqdm==4.66.6
hydra-core==1.3.2
omegaconf==2.3.0
graphviz==0.20.3
pydotplus==2.0.2
ucimlrepo
seaborn==0.13.2